Edit: - AuthorController > method: editAuthor - AuthorService > method: update, findAuthorById - fix and test - IAuthorService > method: update, findAuthorById - authorAdd.jsp > required="true" - authorEdit.jsp > hiddendiv with authorIdToEdit, required="true" - pom.xml - delete unuse dependency

+++ src/main/java/pl/dm/libraryprojectdm/service/IAuthorService.java
@@ -9,7 +9,7 @@ public interface IAuthorService {
 
     List<Author> findAll();
 
-    Author findAuthorById(int id);
+    Author findAuthorById(Long id);
 
     void save(Author author);
